refactor(BottomSheet): 카테고리 및 거리 선택 박스 컴포넌트 분리
