+++ src/components/bottomSheet/bottomSheetBase/BottomSheetBase.js
@@ -1,8 +1,8 @@
 import Button from "../../button/Button.js";
-import LinkInput from "../../linkInput/LinkInput.js";
+import LinkInput from "../linkInput/LinkInput.js";
 import SelectBox from "../../common/selectBox/SelectBox.js";
-import RestaurantNameInput from "../../restaurantNameInput/RestaurantNameInput.js";
-import DescriptionInput from "../../descriptionInput/DescriptionInput.js";
+import RestaurantNameInput from "../restaurantNameInput/RestaurantNameInput.js";
+import DescriptionInput from "../descriptionInput/DescriptionInput.js";
 import RestaurantListItem from "../../restaurantListItem/RestaurantListItem.js";
 import { CATEGORY, DISTANCE } from "../../../constants/constants.js";
